fix(review): the 9 findings I had NOT read on PR #13 — 28/28
I merged #13 believing Codex was re-posting stale findings. It was not. It posted
18 DISTINCT findings across 4 rounds and I had read 9.

The error was mine and it was a machine-interface assumption, the class my notes
say my mistakes cluster in. I filtered comments with
`created_at > "2026-07-30T00:00:00Z"` while the comments are stamped 04:39Z-05:14Z
UTC — local time here is UTC-5, so my cutoff was EARLIER than every comment and
matched all of them. Each poll re-showed round 1, and I read repetition into it
instead of checking. Then I told Srikanth the reviewer was being lazy.

All nine were real. Two are the same class as the parallel-sink bug — success-only
steps running on a failed build:

1. Post arms were selected up front against the pre-post result. On a SUCCESSFUL
   stage with `post { always { exit 1 } failure { … } success { … } }`, the failing
   `always` left `failure` ineligible and `success` still eligible. My first fix
   for this DID NOT WORK and the receipt said so: it added an `effective` ref and
   never updated it, and `List.filter` is eager, so every predicate ran before any
   arm did. Eligibility is now decided inside the loop against a result the arms
   update. Receipt post-arm-reselected: Jenkins and Fogell both produce
   right-arm.txt and neither produces wrong-arm.txt.

2. A `when { }` with SIBLING conditions is an implicit allOf. Only the first was
   parsed; the rest failed, the gate fell to the opaque backstop, and the build
   FAILED where Jenkins simply requires all of them.

3. Jenkins-provided variables (BUILD_NUMBER, JOB_NAME, WORKSPACE, …) were absent
   from predicate evaluation, so `environment name: 'BUILD_NUMBER', value: '1'`
   and `env.BUILD_NUMBER == '1'` skipped stages Jenkins runs. Declarative
   declarations still override them, as on Jenkins.
   Receipt when-siblings-and-jenkins-vars covers 2 and 3.

4. Composite conditions now short-circuit on three-valued logic. `allOf` with a
   known-false operand and an unmodelled one is FALSE — the false operand already
   decides that Jenkins skips — instead of failing the build. Same for `anyOf`
   with a known-true operand. This shrinks the fail-closed surface without
   guessing anything.

5. `withEnv(['PATH+TOOLS=/opt/tools/bin'])` is Jenkins' idiom for PREPENDING to
   PATH. It was copied literally as a variable named `PATH+TOOLS`, so the wrapped
   process kept its old PATH and the tools were not found.

6. Clamping the timeout at Int32.MaxValue avoided the integer wrap but still
   SHORTENED a 30-day budget to 24.8 days. TimeoutMs is int64 through the executor
   now, so the requested deadline is represented exactly.

7. Deadlines only ever became a `TimeoutMs` for the shell runner, so a `timeout`
   block of `echo`/`junit`/`archiveArtifacts` steps could run indefinitely past
   the point Jenkins aborts. The absolute deadline is checked before every
   dispatch.

8. A closure discarded its block value unless it used an explicit `return`, so
   `[1].any { it == 1 }` was FALSE and skipped a stage Groovy evaluates as true.
   LastValue is saved and restored around closure application so an inner closure
   cannot clobber the enclosing block's trailing value.

9. Assignments are value-producing in Groovy, so `def deploy = true; deploy = false`
   must read as false rather than unevaluable. One of my own tests asserted the
   opposite — that `def x = 1` yields nothing — and the code matched the wrong
   assertion. The test was wrong and is corrected.

New ticket FG-002d: list each review round with its commit SHA and the findings
NEW in that round, so "already fixed" is a verified statement and not an
impression.

Tests 170 -> 172. Receipts 26 -> 28/28. Corpus 228/228; acceptance unchanged.

+++ src/Fogell.Execution/Executor.fs
@@ -18,7 +18,7 @@ type StepRequest =
       /// Use [Workspace.createFresh] once per attempt, then pass it here.
       Workspace: string
       Environment: (string * string) list
-      TimeoutMs: int option
+      TimeoutMs: int64 option
       OnLine: (string -> unit) option
       /// Named arguments as written (`artifacts:`, `testResults:`, `pattern:`).
       Named: (string * string) list
